Fix mistakes.

+++ src/File_out.c
@@ -6,7 +6,7 @@
 void save_map(const Map *m, FILE *f) {
   fprintf(f, "%hhu;%hhu;%s", m->w, m->h, m->symbols);
 
-  for (int i = 0; i < m->w * m->h; i++) {
+  for (uint16_t i = 0; i < m->w * m->h; i++) {
     fputc('0' + m->cells[i], f);
   }
 
@@ -14,7 +14,7 @@ void save_map(const Map *m, FILE *f) {
 }
 
 void print_map(const Map *m, FILE *f) {
-  for (int i = 0; i < m->w * m->h; i++) {
+  for (uint16_t i = 0; i < m->w * m->h; i++) {
     fputc(m->symbols[m->cells[i]], f);
 
     if ((i + 1) % m->w == 0) {
